# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/login": {
            "post": {
                "description": "用户（承兑商）登录系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台 API"
                ],
                "summary": "登录系统",
                "parameters": [
                    {
                        "description": "Login argument",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.LoginArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（status为success）失败（status为fail）都会返回200",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.LoginRet"
                        }
                    }
                }
            }
        },
...
        "/merchant/login": {
            "post": {
                "description": "用户（承兑商）登录系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "承兑商登录APP",
                "parameters": [
                    {
                        "description": "输入参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.LoginArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.LoginRet"
                        }
                    }
                }
            }
        },
        "/merchant/logout": {
            "post": {
                "description": "承兑商退出登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "承兑商退出登录",
                "parameters": [
                    {
                        "description": "输入参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.AppLogoutArg"
                        }
                    }
                ]
            }
        },
...
        "/merchant/randomcode": {
            "get": {
                "description": "获取随机验证码，通知短信或者邮件发送。这个API在承兑商“重置密码”时使用。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "获取随机验证码，承兑商“重置密码”时使用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "图形验证码",
                        "name": "v_code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "手机号或者邮箱",
                        "name": "account",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.GetRandomCodeRet"
                        }
                    }
                }
            }
        },
        "/merchant/register": {
            "post": {
                "description": "用户（承兑商）注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "承兑商注册APP",
                "parameters": [
                    {
                        "description": "输入参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.RegisterArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.RegisterRet"
                        }
                    }
                }
            }
        },
        "/merchant/resetpassword": {
            "post": {
                "description": "重置承兑商密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "重置承兑商密码",
                "parameters": [
                    {
                        "description": "输入参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.ResetPasswordArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.ResetPasswordRet"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.AppLogoutArg": {
            "type": "object",
            "properties": {
                "uid": {
                    "type": "integer"
                }
            }
        },
...
        "response.GetRandomCodeRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object"
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.LoginArg": {
            "type": "object",
            "required": [
                "account",
                "password"
            ],
            "properties": {
                "account": {
                    "type": "string",
                    "example": "13112345678"
                },
                "password": {
                    "type": "string",
                    "example": "pwd123"
                }
            }
        },
        "response.LoginRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "nickname": {
                            "description": "用户昵称",
                            "type": "string",
                            "example": "老王"
                        },
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        },
                        "user_cert": {
                            "description": "user_cert可以为0/1，分别表示“已认证/未认证”",
                            "type": "integer"
                        },
                        "user_status": {
                            "description": "user_status可以为0/1/2，分别表示“正常/待审核/冻结”",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.RegisterArg": {
            "type": "object",
            "required": [
                "email",
                "password",
                "phone",
                "pic_code",
                "random_code"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "xxx@sina.com"
                },
                "password": {
                    "type": "string",
                    "example": "pwd1234"
                },
                "phone": {
                    "type": "string",
                    "example": "13112345678"
                },
                "pic_code": {
                    "description": "图形验证码",
                    "type": "string",
                    "example": "E87A"
                },
                "random_code": {
                    "description": "随机验证码（通过手机或邮件发送的）",
                    "type": "string",
                    "example": "9823"
                }
            }
        },
        "response.RegisterRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.ResetPasswordArg": {
            "type": "object",
            "required": [
                "account",
                "password",
                "pic_code",
                "random_code"
            ],
            "properties": {
                "account": {
                    "description": "所要重置密码的账号名",
                    "type": "string",
                    "example": "13112345678"
                },
                "password": {
                    "description": "所设置的新密码",
                    "type": "string",
                    "example": "pwd1234"
                },
                "pic_code": {
                    "description": "图形验证码",
                    "type": "string",
                    "example": "E87A"
                },
                "random_code": {
                    "description": "随机验证码（通过手机或邮件发送的）",
                    "type": "string",
                    "example": "9823"
                }
            }
        },
        "response.ResetPasswordRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object"
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}